Refactor: 拆分 DX 与 OpenGL hook 捕获模块
将 DisplayHook 收窄为渲染方法定位和 MinHook 生命周期管理，截图实现拆分到 D3D9、D3D10、D3D11、D3D12 和 OpenGL 独立模块。

新增 DxCaptureCommon 承载 CopyImageData、DXGI 格式归一化和图像格式判断，减少 D3D10/D3D11 捕获路径重复逻辑。

补强 DX hook 读回路径：D3D10/D3D11 对 MSAA 后备缓冲先 ResolveSubresource 再读回，D3D10/D3D11/D3D12 跳过 DXGI_PRESENT_TEST 测试帧。

统一 D3D12 命名，将旧 Dx12Hook 迁移为 D3D12Capture，并移除本地 vendor/d3dx12.h。

构建依赖改用 vcpkg directx-headers 和 Microsoft::DirectX-Headers 官方 target，同时为静态 MinHook triplet 提供 CMake package 查找路径。
