+++ libop/hook/D3D12Capture.h
@@ -1,6 +1,5 @@
-
 #pragma once
-// ref https://github.com/eugen15/directx-present-hook.git
+
 #include <Windows.h>
 
 #include <cstdint>
@@ -12,33 +11,24 @@
 
 namespace op::hook {
 
-class Dx12Hook final {
+HRESULT __stdcall dx12_hkPresent(IDXGISwapChain *thiz, UINT SyncInterval, UINT Flags);
+
+class D3D12Capture final {
   public:
-    static Dx12Hook *Get();
+    static D3D12Capture *Get();
 
-    // This method must be called only once!
-    // There is no additional check inside.
-    // HRESULT Hook();
-
-    // Captures some frames.
     HRESULT CaptureFrames(HWND windowHandleToCapture, std::wstring_view folderToSaveFrames, int maxFrames);
     void CaptureFrame(IDXGISwapChain *swapChain);
 
   private:
-    Dx12Hook();
-    ~Dx12Hook();
+    D3D12Capture();
+    ~D3D12Capture();
 
-    // The command chain offset from the swap chain object pointer.
     std::uintptr_t commandQueueOffset_ = 0;
-
-    // Hook pointers.
     std::uint64_t presentPointer_ = 0;
     std::uint64_t presentTrampoline_ = 0;
 
-    // The resource to read frames from GPU.
     Microsoft::WRL::ComPtr<ID3D12Resource> readbackResource_;
-
-    // Command allocator for a command list to copy the textures.
     Microsoft::WRL::ComPtr<ID3D12CommandAllocator> commandAllocator_;
 
     UINT readbackDataWidth_ = 0;
@@ -47,7 +37,6 @@ class Dx12Hook final {
 
     void *readbackData_ = nullptr;
 
-    // Capture details.
     HWND windowHandleToCapture_ = NULL;
     std::wstring folderToSaveFrames_;
     int frameIndex_ = 0;
