+++ libop/hook/D3D9Capture.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <d3d9.h>
+
+namespace op::hook {
+
+HRESULT __stdcall dx9_hkEndScene(IDirect3DDevice9 *thiz);
+
+} // namespace op::hook
